+++ src/app/lib/widgets/fhir-servers-dlg/fhir-servers-dlg.component.ts
@@ -1,46 +1,57 @@
 /**
  * Dialog to select FHIR server from the list of servers.
  */
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, ElementRef, Input, OnInit, ViewChild} from '@angular/core';
 import {NgbActiveModal, NgbModal} from '@ng-bootstrap/ng-bootstrap';
-import {FhirService} from '../../../services/fhir.service';
+import {FHIRServer, FhirService} from '../../../services/fhir.service';
+import { UserSpecifiedServerDlgComponent } from '../user-specified-server-dlg/user-specified-server-dlg.component';
 
 @Component({
   selector: 'lfb-fhir-servers-dlg',
   template: `
-    <div class="modal-header bg-primary">
-      <h4 class="modal-title text-white">Choose a FHIR server</h4>
-      <button type="button" class="close" aria-label="Close" (click)="dismiss()">
-        <span class="text-white" aria-hidden="true">&times;</span>
-      </button>
-    </div>
-    <div class="modal-body">
-      <table class="table table-striped">
-        <thead>
-        <tr>
-          <th scope="col"></th>
-          <th scope="col">Fhir Server</th>
-          <th scope="col">Description</th>
-        </tr>
-        </thead>
-        <tbody ngbRadioGroup name="fhirServer" [(ngModel)]="selectedServer">
-        <tr *ngFor="let fhirServer of fhirServerList; index as i">
-          <th scope="row">
-            <label ngbButtonLabel><input ngbButton type="radio" [value]="fhirServer"></label>
-          </th>
-          <td>{{fhirServer.displayName}}</td>
-          <td>{{ fhirServer.desc}}</td>
-        </tr>
-        </tbody>
-      </table>
-    </div>
-    <div class="modal-footer btn-group-sm">
-      <button type="button" class="btn btn-primary" (click)="continue()">Continue</button>
-      <button type="button" class="btn btn-primary" (click)="dismiss()">Cancel</button>
+    <div role="dialog" aria-labelledby="serverDlgTitle" aria-describedby="ServerListCaption">
+      <div class="modal-header bg-primary">
+        <h4 id="serverDlgTitle" class="modal-title text-white">Choose a FHIR server</h4>
+        <button type="button" class="close" aria-label="Close" (click)="dismiss()">
+          <span class="text-white" aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body">
+        <table class="table table-sm table-striped table-bordered">
+          <caption id="serverListCaption">List of available FHIR servers.</caption>
+          <thead>
+          <tr>
+            <th scope="col"></th>
+            <th scope="col">Fhir Server</th>
+            <th scope="col">FHIR Version</th>
+            <th scope="col">Description</th>
+          </tr>
+          </thead>
+          <tbody ngbRadioGroup name="fhirServer" [(ngModel)]="selectedServer">
+          <tr *ngFor="let fhirServer of fhirServerList; index as i">
+            <th scope="row" class="align-middle">
+              <label ngbButtonLabel class="m-0 p-0"><input [attr.id]="fhirServer.endpoint" ngbButton type="radio" [value]="fhirServer"></label>
+            </th>
+            <td class="align-middle"><label [attr.for]="fhirServer.endpoint">{{fhirServer.endpoint}}</label></td>
+            <td class="align-middle">{{ fhirServer.version}}</td>
+            <td class="align-middle">{{ fhirServer.desc}}</td>
+          </tr>
+          </tbody>
+        </table>
+      </div>
+      <div class="modal-footer btn-group-sm border-0">
+        <button type="button" class="btn btn-primary" (click)="addFHIRServer()">Add your FHIR server...</button>
+        <button type="button" class="btn btn-primary" (click)="continue()">Continue</button>
+        <button type="button" class="btn btn-primary" (click)="dismiss()">Cancel</button>
+      </div>
     </div>
   `,
-  styles: [
-  ]
+  styles: [`
+    caption {
+      caption-side: top;
+      font-size: 130%;
+    }
+  `]
 })
 export class FhirServersDlgComponent implements OnInit {
 
@@ -57,6 +68,20 @@ export class FhirServersDlgComponent implements OnInit {
 
   ngOnInit(): void {
   }
+
+  /**
+   * Button handler to add user specified FHIR server to the list.
+   */
+  addFHIRServer() {
+    const modalRef = this.modalService.open(UserSpecifiedServerDlgComponent);
+    modalRef.result.then((fhirServer: FHIRServer) => {
+      if(!this.fhirService.hasFhirServer(fhirServer.endpoint)) {
+        this.fhirService.addNewFhirServer(fhirServer);
+      }
+      this.selectedServer = this.fhirService.getFhirServer();
+    }, (cancelled) => {});
+  }
+
 
   /**
    * Handler for continue button